backend/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
defusedxml==0.7.1
Django==5.2.6
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
factory_boy==3.3.3
Faker==37.8.0
idna==3.10
oauthlib==3.3.1
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
